Update __init__.py
Hi there
I have solved https://github.com/Kalebu/alright/issues/24 
In the above code, I left everything the same and created a second send_message calling it send_message1. In this I am not using find user with the suffix link as this causes the desktop WhatsApp to open, I rather browse to a number, and if the number does NOT exist I then handle this problem by testing for the Button and or Message text box. If Button, then the number does not exists...